pyproject.toml
[project]
name = "joermungandr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "flax>=0.12",
    "jax>=0.8",
    "optax>=0.2",
    "orbax-checkpoint>=0.11.31",
    "sentencepiece>=0.2.1",
    "tensorboardX>=2",
    "tyro>=1.0.3",
    # Pretrained Qwen-3 reranker dependencies
    "transformers>=4.51.0",
    "huggingface-hub>=0.20",
    "peft>=0.7",
    "torch>=2.0",
    "accelerate>=0.25",
]
